% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{constants-iSEEfier}
\alias{constants-iSEEfier}
\alias{iSEE_panel_colors}
\title{Constant values used throughout iSEEfier}
\format{
An object of class \code{character} of length 17.
}
\usage{
iSEE_panel_colors
}
\description{
Constant values used throughout iSEEfier
}
\section{Panel colors}{

\itemize{
\item color values (as string character or hex value) for the panels included by
default in \code{iSEE} and \code{iSEEu}
}
}

\keyword{datasets}
