% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glue_initials}
\alias{glue_initials}
\title{Title}
\usage{
glue_initials(
  ...,
  remove_duplicate_panels = TRUE,
  verbose = TRUE,
  custom_panels_allowed = NULL
)
}
\arguments{
\item{remove_duplicate_panels}{}
}
\description{
Title
}
