% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfier-pkg.R
\docType{package}
\name{iSEEfier-pkg}
\alias{iSEEfier-package}
\alias{iSEEfier-pkg}
\title{iSEEfier: a very convenient way to fire up your iSEE instance}
\description{
iSEEfier provides a set of functionality to quickly create, inspect, and
combine \code{initial} configuration objects.
These can be conveniently passed to the function call to launch \code{iSEE()}
in this manner.
This currently works with the sets of panels provided by the \code{iSEE} and \code{iSEEu}
packages, but can be extended to accommodate the usage of any custom panel
(e.g. from \code{iSEEde}, \code{iSEEpathways}, or any panel developed independently by the
user).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NajlaAbassi/iSEEfier}
  \item Report bugs at \url{https://github.com/NajlaAbassi/iSEEfier/issues}
}

}
\author{
\strong{Maintainer}: Najla Abassi \email{abassi.nejla96@gmail.com} (\href{https://orcid.org/0000-0001-8357-0938}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
