% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfier.R
\name{iSEEfier}
\alias{iSEEfier}
\title{iSEEfier}
\usage{
iSEEfier(
  sce,
  feature.list,
  reddim.type = "TSNE",
  clusters = colnames(colData(sce))[1],
  conditions = colnames(colData(sce))[1]
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{feature.list}{A character vector containing a list of genes}

\item{reddim.type}{A string vector containing the dimensionality reduction type}

\item{clusters}{A character string containing the name of the clusters/cell-type/state (as listed in the colData of the sce)}

\item{conditions}{A character string of the groups/conditions.. (as it appears in the colData of the sce)}
}
\value{
A list of "Panel" objects specifying the initial state of iSEE instance
}
\description{
iSEEfier
}
\examples{
sce <- scRNAseq::RichardTCellData()
sce <- scuttle::logNormCounts(sce)
sce <- BiocSingular::runPCA(sce)
sce <- scater::runTSNE(sce)
gene_list <- c("ENSMUSG00000026581", "ENSMUSG00000005087", "ENSMUSG00000015437")
cluster <- "stimulus"
condition <- "single cell quality"
initial <- iSEEfier(sce = sce, feature.list = gene_list, clusters = cluster, conditions = condition)
}
